added API

+++ src/main/java/com/HMB/NotesProject/notes/NoteService.java
@@ -3,18 +3,20 @@ package com.HMB.NotesProject.notes;
 import com.HMB.NotesProject.users.User;
 import com.HMB.NotesProject.users.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
-
+@Slf4j
 @Service
 public class NoteService {
 
     private final NoteRepository noteRepository;
     private final UserRepository userRepository;
-
+    @Autowired
     public NoteService(NoteRepository noteRepository, UserRepository userRepository) {
         this.noteRepository = noteRepository;
         this.userRepository = userRepository;
@@ -23,18 +25,15 @@ public class NoteService {
     public Note getNoteById(Integer noteId, Integer userId) {
         Note note = noteRepository.findById(noteId)
                 .orElseThrow(() -> new IllegalArgumentException("Note not found with id: " + noteId));
-
         if (!note.getUser().getId().equals(userId)) {
             throw new IllegalArgumentException("Note with id " + noteId + " does not belong to user with id " + userId);
         }
-
         return note;
     }
 
     public Note createNote(Note note, Integer userId) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("User not found with id: " + userId));
-
         note.setUser(user);
         return noteRepository.save(note);
     }
@@ -49,7 +48,11 @@ public class NoteService {
     }
 
     public void deleteNoteById(Integer noteId, Integer userId) {
-        Note note = getNoteById(noteId, userId);
+        Note note = noteRepository.findById(noteId)
+                .orElseThrow(() -> new IllegalArgumentException("Note not found with id: " + noteId));
+        if (!note.getUser().getId().equals(userId)) {
+            throw new IllegalArgumentException("Note with id " + noteId + " does not belong to user with id " + userId);
+        }
         noteRepository.delete(note);
     }
